Fix to current task gold bug
Gold was being saved to the map and file on logout even if a task was not yet complete causing "This task:" on the infobox to show only the current login session if you log out halfway through a task. Gold will now be saved in a config property to maintain between logins and only be pushed to the map and file after a task is complete or cancelled.